fix: typing for CreateScriptHook
