+++ packages/sdk/src/types/index.ts
@@ -3,3 +3,4 @@ export * from './manifest';
 export * from './stats';
 export * from './snapshot';
 export * from './plugins';
+export * from './hooks';
